Respect JSON content-type when encoding request body.
Normally, when ring.mock.request/body receives a params map, it sets the
request content-type to "x-www-form-urlencoded" and urlencodes the
params; that is, {:a 1 :b 2} becomes "a=1&b=2".

If the content-type of the request is "application/json", it is
appropriate to preserve the content-type, and encode the params map as a
JSON string; that is, {:a 1 :b 2} becomes "{\"a\": 1, \"b\": 2}".

This should make it easier to use ring.mock.request to test JSON APIs.